+++ changelog.md
@@ -8,6 +8,9 @@
 - Fixed planet command not requiring OP
 - Fixed armor values for space suit chest and leggings reversed. (#551)
 - Fixed Etrionic Capacitor not draining when powering jet suit.
+- Fixed scrolling not resting in planet screen when navigating through the menu
+- Fixed planet buttons being clickable outside of the visible scroll area.
+- Fixed solar systems in the planet screen being ordered randomly. It's now sorted alphabetically.
 - Added Russian translations (#509) - MaleSkam
 - Added Dutch translations (#525) - marlester-dev
 - Added Chinese translations (#523) - whyvo1